Quick output parsing
My code's fairly ugly in this commit, but it does appear to work.  It walks through the output and correctly recognizes passing and failing tests.  I'm keeping this in my own branch just in case it turns out I'm going in the wrong direction with my commits.

Also, I don't have tests for my code ... mainly because there aren't any tests for existing code in the file I was editing.  I'm still making this commit, though, 'cause I'm about to head home and want to be able to work on it tonight and tomorrow.