Cleanup and Coverage Checking
Cleaned up some of the gorc code.  I moved global variables and constants to the top of the file because they make more sense to me there, cleaned up some calls to strings.Split that are better handled by a call to strings.Fields, and added support for the 'go test -cover' command with 'gorc coverage'.